Make tests less implementation specific
This updates tests so they are less reliant on data attributes as the
refactoring I've worked on loses a bunch of these, arguably too we
should be checking mostly on the user noticable effects than the
implementation.

I found it super hard to write tests though in this style - maybe I'm
spoilt by Capybara? - so my lack of knowledge there meant I didn't
really enhance them much. There's still all sorts of behaviours we could
check for but I don't know how to write them in a readable way.