Somewhat substantial refactor
In an effort to try the resolve the various niggly issues we were
getting with each tweak I took the bold, maybe foolish, step of
addressing them with a re-modelling of how this component works.

Previously this tried to store all state within the MillerColumnsElement
element itself through data attributes and selected states. My
assessment was that this becomes very difficult to understand and reason
with and is hence prone to bugs. I took the step to instead model the
taxonomy separately and have both elements update their visual display
based on that.

This is a bit of a rough and ready just passes tests implementation now
there are some methods that are a bit grim and I had a miserable time
trying to work with flow with the DOM model - that's probably evident in
some of the more verbose parts of the code.