Rename <govuk-breadcrumbs />
This is renamed to <govuk-miller-columns-selected /> as
govuk-breadcrumbs seems to be a bit too assuming for this to have as the
element is not usable outside of a pairing with <govuk-miller-columns
/>.

I felt it was most approriate to give it a name that reflected this
connected and then agonised over the suffix to give it before eventually
giving it 'selected'. Maybe someone with more imagination than I can
name it better in the future.