Polish up the JS a bit
This does some refactoring to clean the code up a bit, reduce number of
methods/duplication and to try make the code just easier to understand.

As the code is quite tricky to understand I've been quite liberal with
commenting to try help the next person so hopefully I'm not the only
person who ever understands how this works.