Rename HTML elements
This renamed <govuk-miller-columns /> to <miller-columns /> and
<govuk-miller-columns-selected /> to <miller-columns-selected />

This is because the govuk prefix to custom elements is somewhat
controversial within GDS. In the realm of CSS a govuk prefix applies to
the design system, pub applies to GOV.UK publishing and app is an
individual application.

The simplest approach to resolving this seems to be to just drop a
prefix.