Provide API method to access selected topic names
This provides a method on the <miller-column-selected> element of
selectedTopicNames which returns lists of the topic names that are
selected.

This is intended to be used by Google Tag Manager so that tags that are
selected can be sent as names rather than ids.

Example:

If we have topics like so:

```
Grandparent of A | Parent of A | A
Parent of B | B | Child of B
Grandparent of C | Parent of C | C
```

and topics B and A are selected

a call to selectedTopicNames would return

```
[
  ['Grandparent of A', 'Parent of A', 'A'],
  ['Parent of B', 'B']
]
```