Remove checkboxes from focus order
If we set the `tabindex` to `-1` in markup (as we did in the examples presented
in this repository) the users that have JavaScript disabled will not be able to
reach them. Instead, if we set them via JavaScript (which is what this PR
proposes) we have the miller-columns items added to the focus order and the
checkboxes removed from the focus order.