Remove component including from the distributed css file
It is expected that an application including this widget already has
govuk-frontend installed. Importing components in the SCSS means that,
most likely, the application ends up with the components output in CSS
twice. This can cause some unpredictable behaviour if anything is
relying on CSS specificity.

This also updates the govuk-frontend including to make use of the new
including of dependencies via the base file.