Stop including core/typography in miller-columns.scss
Importing this file causes side-effects to other apps because it outputs
a variety of classes (govuk-heading-*, govuk-body, etc). This can cause
problems as these rules may be output at a different point in the
ordering of the file and this can cause problems in specificity. This
also causes apps to have duplicate CSS output, and - should
govuk-frontend versions differ between apps - CSS that potentially
overrides newer versions with older versions.

The approach that seems cleanest is to stop the CSS file in this app
outputting any govuk-* class, which is a somewhat manual effort.

In order to cater for this I had to remove the usage of the
%govuk-header-s mixin. The file that introduces this mixin also causes
the typography classes to be output. Instead I've moved in code that
achieves the same style.

An example of where this caused a problem is Content Publisher. The
later introduction of extra `.govuk-heading-m` classes was causing
specificity problems with `<h3 class="govuk-heading-m
gem-c-summary-list__group-title">Content</h3>` where the second output
of `.govuk-heading-m` was overriding the `display: inline-block` defined
on `.gem-c-summary-list__group-title`.