Update eslint-plugin-relay dependency
The build on main is currently failing. This is due to a conflict in the
package.json.lockfile caused by eslint-plugin-relay requiring a version
of GraphQL that is no longer supported by NPM.

Adding the latest version of `eslint-plugin-relay` resolves the issue.

This will allow us to continue to update the dependencies.